Setting up map and removing unneeded files
I added the map functionality and resized the screen and map for 64x64
tiles. If the screen is still too large on Macs, just let me know and
we'll find a better size.

The joke:
Why was the little girl crying?
Because there was a frog stapled to her face