Implemented the in game map
Hitting pause in the main gameplay now causes the map to appear with the
locations of the trees, the player, and the exclamation point. This
calls for celebration, so my joke is:

A man was driving in the middle of nowhere down a secluded country road
far from any cities. He got a flat tire, and got out to walk for
help.After walking for some time, he came to a small stone monastery. He
knocked on the door and roused the monks. "I've got a flat tire. Can I
use your phone?" He asked.

The monks said they were sorry, but they did not have a phone. "If you
stay tonight, you can get a ride on our wagon into town tomorrow," they
said. So the man stayed the night, and they put him in a small room in
the monastery.

In the middle of the night, the man was awakened suddenly by a noise.
Not just any noise, but the loudest, most wonderful, most terrifying,
most hair-raising noise ever.

He sat there, his heart beating for a few minutes, and he heard it
again!Getting out of bed, he went running in the direction of the noise.
It came again, making the hair on the back of his neck rise and his skin
crawl. Finally, he came to a large door where the head monk was
standing. The door was at least 15 feet tall, and made of solid-looking
wood and metal. It had chains and bars and locks and a deadbolt on it,
and was the most formidable door the man had ever seen.

"What was that sound?" He asked. "What made it? Is it behind that door?"

The head monk shook his head. "I'm sorry," he said. "I can't tell you;
you're not a monk."

As the man turned away, he heard the noise again. "You have to tell me
what it is," he begged.

"I'm sorry, I can't tell you, you're not a monk," said the monk.

The man tried to sleep, but couldn't get the noise out of his head. In
the morning, as he was getting ready to leave, he heard the sound again.
It made his ears ring and his mind whirl."Please tell me what made that
sound," he said.

But the monks wouldn't. "I'm sorry, you're not a monk" was all they
said.

The man left, and eventually got his car fixed and went back to his
life. But he couldn't get the sound out of his mind. After a few months,
he got in his car and drove and drove until he found the monastery
again. He got out of his car and found the head monk. "I can't forget
that sound from that night I was here. Please, please please tell me
what made that sound." The head monk just shook his head.

"I can't tell you; you're not a monk," he said."Then tell me how I can
become a monk," the man said.

The head monk said "It's very difficult. Are you sure you want to do
this?"The man said "I've got to. I have to know what made that
sound."The head monk said, "To join us, you have to perform several
tasks. Your first task is to count all of the stars visible in the sky."

The man thought about how hard that would be, but he had to know what
made that sound. He sat up every night for a year, counting the stars
over and over until he was sure how many stars were visible in the sky.
He went to the head monk and told him, and the monk nodded.

"Very good. Your next task is to count all of the grains of sand on the
beaches around the world."The man knew this would be even harder, but he
could not get the noise out of his head. He had to know what, what kind
of animal, could make that terrible horrible mind-bending sound. So he
left on his journeys. He crawled the length and breadth of every beach
in the world, counting the grains of sand, and he returned to the
monastery years later.The head monk heard his answer and nodded.

"Excellent. You are almost done. Your final task is to climb to the peak
of the highest mountain in the world, and see yourself in relation to
the rest of creation." And the man knew this would be hard, but he
outfitted himself, and he went to the highest mountain in the world, and
he climbed to the top, and returned months later, older and wiser and
more tired than years before when he had first heard the noise, the
noise that would not leave his mind and that echoed in his every waking
thought.He returned, and the head monk saw that he was wiser, and said
"At last, you are a monk. Come with me."

And they walked through the monastery, its twisting and turning halls,
and as they went the man heard the noise again, over and over, and he
was no longer sure if it was the noise or merely his memory of it.And
finally, finally, he stood in front of the door and the head monk opened
it up, and the man saw what had made the noise.

But, I can't tell you what it was. You're not a monk.