Spinning water token
Q. Why did the monkey fall out of the tree?
A. It was dead.
Q. Why did the second monkey fall out of the tree?
A. It was stapled to the first monkey.
Q. Why did the third monkey fall out of the tree?
A. It thought it was a game.
Q. Why did the refrigerator fall out of the tree?
A. It couldn't climb because it had no arms.
Q. Why did the little girl fall off her bike?
A. She was hit by three monkeys and a refrigerator.